requirements.txt
# Core dependencies
streamlit==1.31.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9

# AI and ML
cohere==4.32
crewai==0.14.0
langchain>=0.1.4,<0.2.0
langchain-openai==0.0.5
duckduckgo-search==4.1.1
openai>=1.10.0,<2.0.0
